Drafted a document for withdrawals for capella
Hopefully this is a good start for documenting the capella upgrade and what users need to know.

I referenced ethdo, which is a tool we don't maintain, but it's also the only tool i know of at the moment that has this functionality...

The capella upgrade is not set to release on a specific date yet, but this kind of concept document will start being useful very soon.

ref. https://github.com/ConsenSys/teku/issues/5680 and https://github.com/ConsenSys/teku/issues/6482