Updated default data-storage-mode to minimal
Reflecting the changes from teku #7392.

The story is a little complicated because of the historic treatment of existing databases to default to 'prune' so they don't lose data, but the TL/DR; for new databases is they will default to minimal from now on, so reflected that in the documentation.