Added the network option for voluntary-exit command
teku ticket 7489 had an issue where this option could not be used to work around an issue in the remote spec, so it was added to voluntary-exit.

It's almost the same as everywhere else, but its not defaulted to mainnet, as by default is preferred to read from the beacon-api. Also auto won't really work in this circumstance I would expect, because that's what we were trying to do by loading from the beacon-api.